+++ src/peer.rs
@@ -189,7 +189,7 @@ impl<S: Socket> Peer<S> {
     ///
     /// Returns [`BadParameter`] if `mtu` is greater than [`PROTOCOL_MAXIMUM_MTU`] or less than
     /// [`PROTOCOL_MINIMUM_MTU`].
-    pub fn set_mtu(&self, mtu: u16) -> Result<(), BadParameter> {
+    pub fn set_mtu(&mut self, mtu: u16) -> Result<(), BadParameter> {
         if mtu > PROTOCOL_MAXIMUM_MTU as u16 || mtu < PROTOCOL_MINIMUM_MTU as u16 {
             return Err(BadParameter {
                 method: "Peer::set_mtu",
@@ -305,6 +305,8 @@ impl<S: Socket> Peer<S> {
     /// Address of the remote peer, or [`None`] if this peer has never been connected.
     ///
     /// If the peer has disconnected, the previously connected peer's address will be returned.
+    /// 
+    /// This value can be safely unwrapped if a peer was obtained from [`Event`](`crate::Event`).
     #[must_use]
     pub fn address(&self) -> Option<S::Address> {
         unsafe { (*self.0).address.assume_init_ref().clone() }